+++ models/columns.go
@@ -8,14 +8,14 @@ func (c *columnManager) Create(column *Column) error {
 	return DB.Create(&column).Error
 }
 
-func (c *columnManager) Update(column *Column, id uint) error {
+func (c *columnManager) Update(column *Column, id uint64) error {
 	return DB.Model(Column{}).
 		Where("id = ?", id).
 		Updates(column).
 		Error
 }
 
-func (c *columnManager) GetById(id uint) (Column, error) {
+func (c *columnManager) GetById(id uint64) (Column, error) {
 	var column Column
 	err := DB.First(&column, id).Error
 	return column, err
